Remove DetectedPlagiarismVetoData
